feat : 게시글 생성, 수정, 삭제시 작성자 지정 및 확인

+++ src/main/java/cafeboard/Post/PostService.java
@@ -3,10 +3,13 @@ package cafeboard.Post;
 import cafeboard.Board.Board;
 import cafeboard.Board.BoardRepository;
 import cafeboard.Comment.CommentRepository;
+import cafeboard.Member.JwtProvider;
+import cafeboard.Member.Member;
 import cafeboard.Member.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class PostService {
@@ -15,24 +18,32 @@ public class PostService {
     private final BoardRepository boardRepository;
     private final CommentRepository commentRepository;
     private final MemberRepository memberRepository;
+    private final JwtProvider jwtProvider;
 
-    public PostService(PostRepository postRepository, CommentRepository commentRepository, BoardRepository boardRepository, CommentRepository commentRepository1, MemberRepository memberRepository) {
+    public PostService(PostRepository postRepository, CommentRepository commentRepository, BoardRepository boardRepository, CommentRepository commentRepository1, MemberRepository memberRepository, JwtProvider jwtProvider) {
         this.postRepository = postRepository;
         this.boardRepository = boardRepository;
         this.commentRepository = commentRepository1;
         this.memberRepository = memberRepository;
+        this.jwtProvider = jwtProvider;
     }
 
     //게시글 생성
-    public Post create (CreatePostRequest createRequest){
+    public Post create (CreatePostRequest createRequest, String token){
         Board board = boardRepository.findById(createRequest.boardId()).orElseThrow(
                 () -> new IllegalArgumentException("존재하지 않는 게시판입니다. 게시글을 생성할 수 없습니다.")
         );
 
+        String username = jwtProvider.getSubject(token);
+
+        Member writer = memberRepository.findByUsername(username).orElseThrow(
+                () -> new IllegalArgumentException("유효하지 않은 사용자입니다. 게시글을 생성할 수 없습니다.")
+        );
 
         Post post = new Post(createRequest.postTitle(),
                 createRequest.postContent(),
-                board);
+                board,
+                writer);
 
         return postRepository.save(post);
     }
@@ -56,10 +67,15 @@ public class PostService {
 
     //게시글 수정
     @Transactional
-    public PostResponse update(Long postId, PostRequest request){
+    public PostResponse update(Long postId, PostRequest request, String token){
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new IllegalArgumentException("수정할 게시글을 찾을 수 없습니다.")
         );
+
+        String username = jwtProvider.getSubject(token);
+        if(!post.getWriter().getUsername().equals(username)){
+            throw new NoSuchElementException("게시물 작성자만 수정할 수 있습니다.");
+        }
 
         post.setPostTitle(request.postTitle());
         post.setContent(request.postContent());
@@ -71,10 +87,16 @@ public class PostService {
 
     // 게시글삭제
     @Transactional
-    public void delete(Long postId){
+    public void delete(Long postId, String token){
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new IllegalArgumentException("삭제할 게시글을 찾을 수 없습니다.")
         );
+
+        String username = jwtProvider.getSubject(token);
+        if(!post.getWriter().getUsername().equals(username)){
+            throw new NoSuchElementException("게시물 작성자만 삭제할 수 있습니다.");
+        }
+
 
         postRepository.delete(post);
     }
